refactor: moves vco related sources to their own folder
