+++ src/virtual-code-owners/__fixtures__/corpus/empty-teams/000-empty.yml
@@ -1,0 +1,1 @@
+[{ type: "empty", line: 1, raw: "" }]
