+++ src/virtual-code-owners/__fixtures__/corpus/empty-teams/001-comment.txt
@@ -1,0 +1,1 @@
+# a comment
