+++ src/virtual-code-owners/__fixtures__/corpus/empty-teams/001-comment.yml
@@ -1,0 +1,1 @@
+[{ type: "comment", line: 1, raw: "# a comment" }]
