+++ src/virtual-code-owners/__fixtures__/corpus/empty-teams/002-ignorable-comment.txt
@@ -1,0 +1,1 @@
+#! an ignorable comment
