+++ src/virtual-code-owners/__fixtures__/corpus/empty-teams/002-ignorable-comment.yml
@@ -1,0 +1,1 @@
+[{ type: "ignorable-comment", line: 1, raw: "#! an ignorable comment" }]
