+++ src/virtual-code-owners/__fixtures__/corpus/empty-teams/003-unparsable-line.txt
@@ -1,0 +1,1 @@
+this-is-not-a-valid-rule
