+++ src/virtual-code-owners/__fixtures__/corpus/empty-teams/003-unparsable-line.yml
@@ -1,0 +1,1 @@
+[{ type: "unknown", line: 1, raw: "this-is-not-a-valid-rule" }]
