+++ src/virtual-code-owners/__fixtures__/corpus/empty-teams/010-rule-other-user.txt
@@ -1,0 +1,1 @@
+* @everyone
