+++ src/virtual-code-owners/__fixtures__/corpus/empty-teams/011-rule-e-mail-user.txt
@@ -1,0 +1,1 @@
+* jan@example.com
