+++ src/virtual-code-owners/__fixtures__/corpus/empty-teams/011-rule-e-mail-user.yml
@@ -1,0 +1,19 @@
+[
+  {
+    type: "rule",
+    line: 1,
+    filesPattern: "*",
+    spaces: " ",
+    users:
+      [
+        {
+          type: "e-mail-address",
+          raw: "jan@example.com",
+          bareName: "jan@example.com",
+          userNumberWithinLine: 1,
+        },
+      ],
+    inlineComment: "",
+    raw: "* jan@example.com",
+  },
+]
