+++ src/virtual-code-owners/__fixtures__/corpus/empty-teams/012-rule-invalid-user.txt
@@ -1,0 +1,1 @@
+* username-does-not-start-with-an-at
