+++ src/virtual-code-owners/__fixtures__/corpus/empty-teams/012-rule-invalid-user.yml
@@ -1,0 +1,19 @@
+[
+  {
+    type: "rule",
+    line: 1,
+    filesPattern: "*",
+    spaces: " ",
+    users:
+      [
+        {
+          type: "invalid",
+          raw: "username-does-not-start-with-an-at",
+          bareName: "username-does-not-start-with-an-at",
+          userNumberWithinLine: 1,
+        },
+      ],
+    inlineComment: "",
+    raw: "* username-does-not-start-with-an-at",
+  },
+]
