+++ src/virtual-code-owners/__fixtures__/corpus/empty-teams/013-rule-with-inline-comment.txt
@@ -1,0 +1,1 @@
+* @everyone   # inline comment
