+++ src/virtual-code-owners/__fixtures__/corpus/empty-teams/014-rule-with-non-space-sepped-inline-comment.txt
@@ -1,0 +1,1 @@
+* @everyone# inline comment
