+++ src/virtual-code-owners/__fixtures__/corpus/teams/020-rule-virtual-team.txt
@@ -1,0 +1,1 @@
+* @baarden
