+++ src/virtual-code-owners/__fixtures__/corpus/teams/021-rule-multiple-teams.txt
@@ -1,0 +1,1 @@
+* @baarden @leren-lezen @real-team jet@example.com
