+++ src/virtual-code-owners/__mocks__/erroneous-virtual-codeowners.txt
@@ -1,0 +1,34 @@
+# The next line is just empty - _that's OK_!
+
+
+# This next line just seemingly contains a file pattern, 
+# but no user or team names. That's quite not OK
+file/pattern/     
+
+* @cloud-heroes-all
+
+# admin & ci stuff => transversal
+
+.github/                @ch/transversal
+
+# This line is hunky dory, except some of the user/ team names don't
+# start with @ or aren't e-mail addresses
+file/pattern @aap @noot mies @wim zus@example.com jet
+
+# generic stuff
+
+apps/framework/         @ch/transversal
+apps/ux-portal/         @ch/ux @ch/transversal
+libs/components/        @ch/ux
+
+# specific functionality
+
+libs/ubc-sales/         @ch/sales
+libs/ubc-after-sales/   @ch/after-sales
+libs/ubc-pre-sales/     @ch/pre-sales
+libs/ubc-refund/        @ch/sales @ch/after-sales
+libs/ubc-baarden/       @ch/mannen-met-baarden
+#!
+#! The VIRTUAL-CODEOWNERS.txt file has the .txt extension because untamed vscode
+#! installations might otherwise interpret it as markdown and auto-correct
+#! and auto-format it as such
