+++ src/virtual-code-owners/parse.test.ts
@@ -3,7 +3,7 @@ import { readFileSync, readdirSync } from "node:fs";
 import { extname, join } from "node:path";
 import { describe, it } from "node:test";
 import { parse as parseYaml } from "yaml";
-import { parse } from "./parse-virtual-code-owners.js";
+import { parse } from "./parse.js";
 
 function relEmpty(pFileName) {
   return new URL(
